Various optimisations
tiles now delete sides that will never be in view. This includes the
bottom side because I don't care if people see that mess.

Drop shadow uses a projector instead of an extra light in the scene.

Imported some assets for projections.

Out of bounds boundaries are now based on map's height instead of just a
hardcoded 1000